Changed the share button to the same color as the message button. Updated the data received from facebook for the age_segment and gender with the new permissions. Changed the rating view to show not based on time but instead based on when the user interacts with the ui such as send button, refresh images and texts.
Changed texts for idées des messages to a new updated texts for notifications.

Sometimes the pulse circle would show even the flame icon would not, this has been changed.

Add a selfie or images from the camera roll, need to know better if the images should be refreshed after the user has pressed a button or based on a completion handler so that they see the new images.